create io utils file
